+++ TelegramBotAPI/src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/requests/stickers/AddAnimatedStickerToSet.kt
@@ -2,6 +2,7 @@ package com.github.insanusmokrassar.TelegramBotAPI.requests.stickers
 
 import com.github.insanusmokrassar.TelegramBotAPI.requests.abstracts.*
 import com.github.insanusmokrassar.TelegramBotAPI.requests.common.CommonMultipartFileRequest
+import com.github.insanusmokrassar.TelegramBotAPI.requests.stickers.abstracts.StandardStickerSetAction
 import com.github.insanusmokrassar.TelegramBotAPI.requests.stickers.abstracts.StickerSetAction
 import com.github.insanusmokrassar.TelegramBotAPI.types.*
 import com.github.insanusmokrassar.TelegramBotAPI.types.stickers.MaskPosition
@@ -35,7 +36,7 @@ data class AddAnimatedStickerToSet internal constructor(
     val sticker: FileId? = null,
     @SerialName(maskPositionField)
     override val maskPosition: MaskPosition? = null
-) : StickerSetAction {
+) : StandardStickerSetAction {
     init {
         if(emojis.isEmpty()) {
             throw IllegalArgumentException("Emojis must not be empty")